+++ sdl2/test/sdl2ext_font_test.py
@@ -3,7 +3,8 @@ import pytest
 from sdl2 import ext as sdl2ext
 from sdl2.ext.compat import byteify
 from sdl2.ext.pixelaccess import pixels2d
-from sdl2 import surface
+from sdl2.ext.surface import _create_surface
+from sdl2 import surface, pixels
 
 _HASSDLTTF = True
 try:
@@ -56,9 +57,17 @@ class TestSDL2ExtFont(object):
         font = sdl2ext.BitmapFont(sf, (32, 32), FONTMAP)
         assert font.size == (32, 32)
 
+        # Try loading a font directly from a .bmp
+        font = sdl2ext.BitmapFont(fontpath, (32, 32), FONTMAP)
+        assert font.size == (32, 32)
+
+        # Test use of default fontmap and inferred character size
+        font = sdl2ext.BitmapFont(sf)
+        assert font.size == (32, 32)
+
         # Try invalid surface type
         with pytest.raises(TypeError):
-            font = sdl2ext.BitmapFont("hello", (32, 32), FONTMAP)
+            sdl2ext.BitmapFont([], (32, 32), FONTMAP)
 
     def test_BitmapFont_render(self):
         # Initialize font and BitmapFont for tests
@@ -76,41 +85,116 @@ class TestSDL2ExtFont(object):
         with pytest.raises(ValueError):
             font.render("this_should_fail")
 
-    def test_BitmapFont_render_on(self):
-        np = pytest.importorskip("numpy", reason="numpy module is not available")
-        # Initialize font, surface, and BitmapFont for tests
+    def test_BitmapFont_render_text(self):
+        # Initialize BitmapFont and dummy RGB888 surface for tests
         fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
-        sf = surface.SDL_LoadBMP(fontpath)
-        font = sdl2ext.BitmapFont(sf.contents, (32, 32), FONTMAP)
+        font = sdl2ext.BitmapFont(fontpath)
+        rgb_surf = _create_surface(size=(320, 256), fmt="RGB888")
+        font_rgb = sdl2ext.BitmapFont(rgb_surf)
 
         # Try rendering some text
-        target = surface.SDL_CreateRGBSurface(0, 32*5, 32, 32, 0, 0, 0, 0)
+        msg = "hello there!"
+        text = font.render_text(msg)
+        assert isinstance(text, surface.SDL_Surface)
+        assert text.w == 32 * len(msg)
+        assert text.h == 32
+        surface.SDL_FreeSurface(text)
+
+        # Try rendering with a different line height
+        text = font.render_text(msg, line_h=40)
+        assert isinstance(text, surface.SDL_Surface)
+        assert text.w == 32 * len(msg)
+        assert text.h == 40
+        surface.SDL_FreeSurface(text)
+
+        # Make sure ARGB converion works
+        text = font_rgb.render_text(msg)
+        assert isinstance(text, surface.SDL_Surface)
+        assert text.w == 32 * len(msg)
+        assert text.h == 32
+        assert text.format.contents.format == pixels.SDL_PIXELFORMAT_ARGB8888
+        surface.SDL_FreeSurface(text)
+
+        # Try rendering without ARGB conversion
+        text = font_rgb.render_text(msg, as_argb=False)
+        assert isinstance(text, surface.SDL_Surface)
+        assert text.w == 32 * len(msg)
+        assert text.h == 32
+        assert text.format.contents.format == pixels.SDL_PIXELFORMAT_RGB888
+        surface.SDL_FreeSurface(text)
+
+        # Test multiline rendering
+        msg = "hello\nthere!"
+        text = font.render_text(msg)
+        assert isinstance(text, surface.SDL_Surface)
+        assert text.w == 32 * 6
+        assert text.h == 64
+        surface.SDL_FreeSurface(text)
+
+        # Test exception for missing glyph
+        with pytest.raises(ValueError):
+            font.render_text("this_should_fail")
+
+    def test_BitmapFont_render_on(self):
+        np = pytest.importorskip("numpy", reason="numpy module is not available")
+        # Initialize BitmapFont for tests
+        fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
+        font = sdl2ext.BitmapFont(fontpath)
+
+        # Try rendering some text
+        target = _create_surface(size=(32*5, 32))
         view = pixels2d(target, transpose=False)
         mid_row = view[16, :].copy()
-        font.render_on(target, "TEST!")
+        outrect = font.render_on(target, "TEST!")
         assert not np.all(mid_row == view[16, :]) # ensure surface changed
+        assert outrect == (0, 0, 32*5, 32)
 
         # Try rendering some text with an offset
-        target2 = surface.SDL_CreateRGBSurface(0, 32*5, 32, 32, 0, 0, 0, 0)
+        target2 = _create_surface(size=(32*5, 32))
         view2 = pixels2d(target2, transpose=False)
-        font.render_on(target2, "TEST!", offset=(5, 0))
+        outrect2 = font.render_on(target2, "TEST!", offset=(5, 0))
         assert not np.all(mid_row == view2[16, :]) # ensure surface changed
         assert not np.all(view[16, :] == view2[16, :]) # ensure offset worked
+        assert outrect2 == (5, 0, 32*5 + 5, 32)
+
+        surface.SDL_FreeSurface(target)
+        surface.SDL_FreeSurface(target2)
 
         # Test exception for missing glyph
         with pytest.raises(ValueError):
             font.render_on(target, "%nope")
 
+    def test_BitmapFont_remap(self):
+        # Initialize BitmapFont for tests
+        fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
+        font = sdl2ext.BitmapFont(fontpath)
+        msg = "hello there!"
+
+        # Remap the l to be narrower and try rendering
+        font.remap("l", 32, 160, 12, 32)
+        text = font.render_text(msg)
+        assert text.w == (32 * len(msg) - 40)
+        assert text.h == 32
+
+        surface.SDL_FreeSurface(text)
+
+        # Test exceptions on bad input
+        with pytest.raises(ValueError):
+            font.remap("hi", 4, 4, 4, 4)
+        with pytest.raises(ValueError):
+            font.remap("h", 10, 32, 32, 0)
+        with pytest.raises(ValueError):
+            font.remap("h", 32, 1000, 32, 32)
+
     def test_BitmapFont_contains(self):
-        sf = surface.SDL_LoadBMP(byteify(RESOURCES.get_path("font.bmp"),
-                                         "utf-8"))
-        assert isinstance(sf.contents, surface.SDL_Surface)
-        font = sdl2ext.BitmapFont(sf, (32, 32), FONTMAP)
-        assert isinstance(font, sdl2ext.BitmapFont)
+        # Initialize BitmapFont for tests
+        fontpath = byteify(RESOURCES.get_path("font.bmp"), "utf-8")
+        font = sdl2ext.BitmapFont(fontpath)
 
         for ch in "abcde12345.-,+":
             assert font.contains(ch)
-        for ch in "äöüß":
+        del font.offsets[" "]
+        for ch in " äöüß":
             assert not font.contains(ch)
 
     def test_BitmapFont_can_render(self):